tests/requirements.txt
playwright==1.58.0
pytest==7.4.3
pytest-html==4.1.1
python-dotenv==1.0.0
